add validation for required values and metadata fields
