+++ awx/ui/client/features/credentials/external-test-modal.partial.html
@@ -2,11 +2,18 @@
     <at-form state="vm.form" autocomplete="off" id="external_test_form">
         <at-input-group col="12" tab="20" state="vm.form.inputs" form-id="external_test"></at-input-group>
         <at-action-group col="12" pos="right">
-            <at-action-button variant="tertiary" ng-click="vm.onClose()">
-		{{::vm.strings.get('CLOSE')}}
+            <at-action-button
+                variant="tertiary"
+                ng-click="vm.onClose()"
+            >
+                {{::vm.strings.get('CLOSE')}}
             </at-action-button>
-            <at-action-button variant="primary" ng-click="vm.onSubmit()">
-		{{::vm.strings.get('RUN')}}
+            <at-action-button
+                variant="primary"
+                ng-click="vm.onSubmit()"
+                ng-disabled="!vm.form.isValid || vm.form.disabled"
+            >
+                {{::vm.strings.get('RUN')}}
             </at-action-button>
         </at-action-group>
     </at-form>
