+++ awx/ui/client/features/credentials/input-source-lookup.partial.html
@@ -37,6 +37,7 @@
             <at-action-button
                 variant="secondary"
                 ng-click="vm.onTest()"
+                ng-disabled="!vm.form.isValid || vm.form.disabled"
                 ng-show="vm.tabs.metadata._active"
             >
                 {{::vm.strings.get('TEST')}}
@@ -50,6 +51,7 @@
             <at-action-button
                 variant="primary"
                 ng-click="vm.onNext()"
+                ng-disabled="!vm.selectedId"
                 ng-show="vm.tabs.credential._active"
             >
                 {{::vm.strings.get('NEXT')}}
@@ -57,6 +59,7 @@
             <at-action-button
                 variant="primary"
                 ng-click="vm.onSelect()"
+                ng-disabled="!vm.form.isValid || vm.form.disabled"
                 ng-show="vm.tabs.metadata._active"
             >
                 {{::vm.strings.get('OK')}}
